pom changed

+++ src/main/java/ru/geekbrains/architectures_and_patterns/entities/Product.java
@@ -9,9 +9,9 @@ public class Product  {
     private int productId;
     private String productName;
     private int productCalories;
-    private float productProteins;
-    private float productFats;
-    private float productCarbs;
+    private long productProteins;
+    private long productFats;
+    private long productCarbs;
 
 
 }
